backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
waitress==3.0.1
psycopg[binary]>=3.1.0
redis==5.0.1
Flask-Session==0.5.0
psutil==5.9.6          
humanize==4.8.0
schedule==1.2.0
psutil==5.9.6
aiohttp==3.9.3
pyOpenSSL==23.2.0
cryptography==41.0.7
gunicorn==21.2.0
psycopg2-binary==2.9.9